requirements.txt
Django==4.0.4
django-split-settings==1.1.0
djangorestframework==3.14.0

python-dotenv==0.21.0